API

GRAPE.GradGenerator

GRAPE.GradVector

GRAPE._bw_prop!

GRAPE._fw_prop!

GRAPE._fw_prop_aux!

GRAPE.GradGenerator — Type

Extended generator for the standard dynamic gradient.

G̃ = GradGenerator(G)

contains the original generator G (a Hamiltonian or Liouvillian) in G̃.G, a vector of control generators $∂G/∂ϵᵢ$ (static operators for linear controls, and functions for non-linear controls) in G̃.control_generators, and the controls in G̃.controls.

For a generator $G = Ĥ = Ĥ₀ + ϵ₁(t) Ĥ₁ + … + ϵₙ(t) Ĥₙ$, this extended generator encodes the block-matrix

\[G̃ = \begin{pmatrix} Ĥ & 0 & \dots & 0 & Ĥ₁ \\ 0 & Ĥ & \dots & 0 & Ĥ₂ \\ \vdots & & \ddots & & \vdots \\ 0 & 0 & \dots & Ĥ & Ĥₙ \\ 0 & 0 & \dots & 0 & Ĥ \end{pmatrix}\]

GRAPE.GradVector — Type

Extended state-vector for the dynamic gradient.

Ψ̃ = GradVector(Ψ, G̃)

for a state Ψ and a GradGenerator G̃ contains the original Ψ in Ψ̃.state and storage for $n$ gradient-states for the n controls in G̃ in Ψ̃.grad_states.

The GradVector conceptually corresponds to a direct-sum (block) column-vector $Ψ̃ = (|Ψ̃₁⟩, |Ψ̃₂⟩, … |Ψ̃ₙ⟩, |Ψ⟩)^T$. With a $G̃$ as in the documentation of GradGenerator, we have

\[G̃ Ψ̃ = \begin{pmatrix} Ĥ |Ψ̃₁⟩ + Ĥ₁|Ψ⟩ \\ \vdots \\ Ĥ |Ψ̃ₙ⟩ + Ĥₙ|Ψ⟩ \\ Ĥ |Ψ⟩ \end{pmatrix}\]

and

\[e^{-i G̃ dt} \begin{pmatrix} 0 \\ \vdots \\ 0 \\ |Ψ⟩ \end{pmatrix} = \begin{pmatrix} \frac{∂}{∂ϵ₁} e^{-i Ĥ dt} |Ψ⟩ \\ \vdots \\ \frac{∂}{∂ϵₙ} e^{-i Ĥ dt} |Ψ⟩ \\ e^{-i Ĥ dt} |Ψ⟩ \end{pmatrix}.\]

GRAPE._bw_prop! — Method

Backwards propagate the states ϕ and store them

GRAPE._fw_prop! — Method

Propagate system forward in time and store states at a constant objective index

GRAPE._fw_prop_aux! — Method

Propagate auxilliary system forward in time and then store the states at a constant objective index