API

GRAPE.GrapeResult

GRAPE.optimize_grape

GRAPE.print_table

QuantumControlBase.optimize

GRAPE.GrapeResult — Type

Result object returned by optimize_grape.

GRAPE.optimize_grape — Method

Optimize a control problem using GRAPE.

result = optimize_grape(problem)

optimizes the given control problem, returning a GrapeResult.

Note

It is recommended to call optimize with method=:GRAPE instead of calling optimize_grape directly.

Keyword arguments that control the optimization are taken from the keyword arguments used in the instantiation of problem.

Required problem keyword arguments

J_T: A function J_T(ϕ, objectives, τ=τ) that evaluates the final time functional from a list ϕ of forward-propagated states and problem.objectives.

gradient: A function gradient!(G, τ, ∇τ) that stores the gradient of J_T in G.

Optional problem keyword arguments

update_hook

info_hook

check_convergence

x_tol

f_tol

g_tol

show_trace

extended_trace

show_every

allow_f_increases

optimizer

prop_method/fw_prop_method/bw_prop_method: The propagation method to use for each objective, see below.

prop_method/fw_prop_method/grad_prop_method: The propagation method to use for the extended gradient vector for each objective, see below.

The propagation method for the forward propagation of each objective is determined by the first available item of the following:

a fw_prop_method keyword argument

a prop_method keyword argument

a property fw_prop_method of the objective

a property prop_method of the objective

the value :auto

The propagation method for the backword propagation is determined similarly, but with bw_prop_method instead of fw_prop_method. The propagation method for the forward propagation of the extended gradient vector for each objective is determined from grad_prop_method, fw_prop_method, prop_method in order of precedence.

GRAPE.print_table — Method

Print optimization progress as a table.

This functions serves as the default info_hook for an optimization with GRAPE.

QuantumControlBase.optimize — Method

opt_result = optimize(problem; method=:GRAPE, kwargs...)

optimizes problem using GRadident Ascent Pulse Engineering (GRAPE), see GRAPE.optimize_grape.